feat: stop native input event, fire custom event on user interaction
