+++ packages/slider/src/vaadin-range-slider.d.ts
@@ -17,6 +17,13 @@ export type RangeSliderChangeEvent = Event & {
 };
 
 /**
+ * Fired when the slider value changes during user interaction.
+ */
+export type RangeSliderInputEvent = Event & {
+  target: RangeSlider;
+};
+
+/**
  * Fired when the `value` property changes.
  */
 export type RangeSliderValueChangedEvent = CustomEvent<{ value: number[] }>;
@@ -27,6 +34,7 @@ export interface RangeSliderCustomEventMap {
 
 export interface RangeSliderEventMap extends HTMLElementEventMap, RangeSliderCustomEventMap {
   change: RangeSliderChangeEvent;
+  input: RangeSliderInputEvent;
 }
 
 /**
@@ -38,6 +46,7 @@ export interface RangeSliderEventMap extends HTMLElementEventMap, RangeSliderCus
  * ```
  *
  * @fires {Event} change - Fired when the user commits a value change.
+ * @fires {Event} input - Fired when the slider value changes during user interaction.
  * @fires {CustomEvent} value-changed - Fired when the `value` property changes.
  */
 declare class RangeSlider extends FieldMixin(SliderMixin(FocusMixin(ThemableMixin(ElementMixin(HTMLElement))))) {
