+++ packages/slider/src/vaadin-slider.js
@@ -26,6 +26,7 @@ import { SliderMixin } from './vaadin-slider-mixin.js';
  * ```
  *
  * @fires {Event} change - Fired when the user commits a value change.
+ * @fires {Event} input - Fired when the slider value changes during user interaction.
  * @fires {CustomEvent} value-changed - Fired when the `value` property changes.
  *
  * @customElement
@@ -209,7 +210,9 @@ class Slider extends FieldMixin(
 
   /** @private */
   __onInput(event) {
+    event.stopPropagation();
     this.__updateValue(event.target.value, 0);
+    this.__dispatchInputEvent();
     this.__commitValue();
   }
