+++ packages/slider/test/typings/range-slider.types.ts
@@ -7,7 +7,12 @@ import type { LabelMixinClass } from '@vaadin/field-base/src/label-mixin.js';
 import type { ValidateMixinClass } from '@vaadin/field-base/src/validate-mixin.js';
 import type { ThemableMixinClass } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
 import type { SliderMixinClass } from '../../src/vaadin-slider-mixin.js';
-import type { RangeSlider, RangeSliderChangeEvent, RangeSliderValueChangedEvent } from '../../vaadin-range-slider.js';
+import type {
+  RangeSlider,
+  RangeSliderChangeEvent,
+  RangeSliderInputEvent,
+  RangeSliderValueChangedEvent,
+} from '../../vaadin-range-slider.js';
 
 const assertType = <TExpected>(actual: TExpected) => actual;
 
@@ -22,6 +27,11 @@ slider.addEventListener('value-changed', (event) => {
 
 slider.addEventListener('change', (event) => {
   assertType<RangeSliderChangeEvent>(event);
+  assertType<RangeSlider>(event.target);
+});
+
+slider.addEventListener('input', (event) => {
+  assertType<RangeSliderInputEvent>(event);
   assertType<RangeSlider>(event.target);
 });
